refactor(gate): the engine is a package; the CLI is its first caller
gate/ was one main package: the render/diff/validate/export engine and
the command-line front end compiled into the same binary, so nothing
else could call the engine. Split them. The engine keeps its name and
directory as an importable package; the CLI moves to gate/cmd/gitops-gate
with identical flags, output and exit codes.

Two seams open on the way through, both for a caller that reads the
cluster inventory live instead of from a checked-in snapshot:

- ParseConfig takes bytes, and no longer requires `clusters` -- that key
  names the snapshot, so the requirement moves to the CLI, the only
  reader of one.
- The export filter globals become an ExportFilter value, and
  InventoryFromSecrets decodes ArgoCD cluster Secrets for both the
  kubectl path and any future in-process reader, so the two can never
  decode the same Secret two different ways.
