+++ gate/topologies_test.go
@@ -1,4 +1,4 @@
-package main
+package gate
 
 import (
 	"fmt"
